+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.12.0-rc.17",
-  genVersion: "2.916.4",
-  userAgent: "speakeasy-sdk/typescript 4.12.0-rc.17 2.916.4 2.0 @orq-ai/node",
+  sdkVersion: "4.12.0-rc.18",
+  genVersion: "2.917.0",
+  userAgent: "speakeasy-sdk/typescript 4.12.0-rc.18 2.917.0 2.0 @orq-ai/node",
 } as const;