+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -412,7 +412,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KXCW9ZZFA0APHPXP6ZZ77GE6"),
+  _id: z.string().default("01KXE47WNVGNKNT4WDCZCS6CX4"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: CreateDatasourceStatus$inboundSchema,