+++ packages/orq-rc/RELEASES.md
@@ -10729,3 +10729,13 @@ Based on:
 - [typescript v4.12.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.17 - packages/orq-rc
+
+## 2026-07-13 16:18:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.0 (2.917.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.18 - packages/orq-rc